--- Content/Weapons/BP_Thrust.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:31be06aaaa06ee9a2d0f553a604bdd7c70a1954c2c30e6d6f1bbb1c1e8632d10
-size 15887
