+++ Source/SF/Weapons/SF_WeaponSkillBase.h
@@ -1,0 +1,43 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Abilities/GameplayAbility.h"
+#include "SF_WeaponSkillBase.generated.h"
+
+
+UCLASS(Abstract) 
+class SF_API UWeaponSkillBase : public UGameplayAbility 
+{
+	GENERATED_BODY()
+
+public:
+	UWeaponSkillBase();
+
+	
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Animation")
+	TObjectPtr<UAnimMontage> MontageToPlay;
+
+protected:
+	
+	 
+	virtual void ActivateAbility(const FGameplayAbilitySpecHandle Handle, const FGameplayAbilityActorInfo* ActorInfo, const FGameplayAbilityActivationInfo ActivationInfo, const FGameplayEventData* TriggerEventData) override;
+
+	
+	 
+	virtual void EndAbility(const FGameplayAbilitySpecHandle Handle, const FGameplayAbilityActorInfo* ActorInfo, const FGameplayAbilityActivationInfo ActivationInfo, bool bReplicateEndAbility, bool bWasCancelled) override;
+
+	UFUNCTION()
+	virtual void OnMontageCompleted();
+
+	
+	UFUNCTION()
+	virtual void OnMontageInterrupted();
+	
+	UFUNCTION()
+	virtual void OnMontageCancelled();
+
+protected:
+	
+	UPROPERTY()
+	TObjectPtr<class UAbilityTask_PlayMontageAndWait> MontageTask;
+};
